:truck: move `getCorrectionForAnswer` usecase to `src/evaluation/`
