+++ api/src/evaluation/domain/usecases/index.js
@@ -2,7 +2,6 @@ import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import { certificationCompletedJobRepository } from '../../../../lib/infrastructure/repositories/jobs/certification-completed-job-repository.js';
-import * as courseRepository from '../../../../src/shared/infrastructure/repositories/course-repository.js';
 import * as certificationEvaluationCandidateRepository from '../../../certification/evaluation/infrastructure/repositories/certification-candidate-repository.js';
 import * as certificationChallengeLiveAlertRepository from '../../../certification/shared/infrastructure/repositories/certification-challenge-live-alert-repository.js';
 import * as campaignRepository from '../../../prescription/campaign/infrastructure/repositories/campaign-repository.js';
@@ -10,12 +9,14 @@ import * as campaignParticipationRepository from '../../../prescription/campaign
 import { participationCompletedJobRepository } from '../../../prescription/campaign-participation/infrastructure/repositories/jobs/participation-completed-job-repository.js';
 import * as targetProfileAdministrationRepository from '../../../prescription/target-profile/infrastructure/repositories/target-profile-administration-repository.js';
 import * as targetProfileRepository from '../../../prescription/target-profile/infrastructure/repositories/target-profile-repository.js';
+import { fromDatasourceObject } from '../../../shared/infrastructure/adapters/solution-adapter.js';
 import * as answerRepository from '../../../shared/infrastructure/repositories/answer-repository.js';
 import * as areaRepository from '../../../shared/infrastructure/repositories/area-repository.js';
 import * as assessmentRepository from '../../../shared/infrastructure/repositories/assessment-repository.js';
 import * as badgeForCalculationRepository from '../../../shared/infrastructure/repositories/badge-for-calculation-repository.js';
 import * as challengeRepository from '../../../shared/infrastructure/repositories/challenge-repository.js';
 import * as competenceRepository from '../../../shared/infrastructure/repositories/competence-repository.js';
+import * as courseRepository from '../../../shared/infrastructure/repositories/course-repository.js';
 import { repositories as injectedSharedRepositories } from '../../../shared/infrastructure/repositories/index.js';
 import * as skillRepository from '../../../shared/infrastructure/repositories/skill-repository.js';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
@@ -51,7 +52,9 @@ const dependencies = {
   algorithmDataFetcherService,
   answerJobRepository,
   certificationCompletedJobRepository,
+  fromDatasourceObject,
   answerRepository,
+  correctionRepository: repositories.correctionRepository,
   areaRepository,
   assessmentRepository,
   autonomousCourseRepository: repositories.autonomousCourseRepository,
